+++ wp/wp-content/plugins/cloudinary-image-management-and-manipulation-in-the-cloud-cdn/php/sync/class-storage.php
@@ -74,6 +74,13 @@ class Storage implements Notice {
 	protected $settings;
 
 	/**
+	 * Hold the Delivery instance.
+	 *
+	 * @var Delivery
+	 */
+	protected $delivery;
+
+	/**
 	 * The delay in seconds before local assets get deleted if Cloudinary only storage.
 	 */
 	const DELETE_DELAY = MINUTE_IN_SECONDS;
@@ -85,6 +92,7 @@ class Storage implements Notice {
 	 */
 	public function __construct( \Cloudinary\Plugin $plugin ) {
 		$this->plugin = $plugin;
+
 		add_action( 'cloudinary_register_sync_types', array( $this, 'setup' ), 20 );
 		// Add sync storage checks.
 		add_filter( 'cloudinary_render_field', array( $this, 'maybe_disable_connect' ), 10, 2 );
@@ -296,25 +304,41 @@ class Storage implements Notice {
 	 * @return bool
 	 */
 	public function validate( $attachment_id ) {
-		$valid = true;
-		if ( 'cld' === $this->settings['offload'] ) {
-			// In cld mode, we want to delay the deletion.
-			$delayed = get_post_meta( $attachment_id, Sync::META_KEYS['delay'], true );
-			$now     = time();
-			if ( empty( $delayed ) ) {
-				update_post_meta( $attachment_id, Sync::META_KEYS['delay'], $now );
+		$valid = false;
+
+		if ( $this->delivery->is_deliverable( $attachment_id ) ) {
+			$valid = true;
+
+			if ( 'cld' === $this->settings['offload'] ) {
+				// In cld mode, we want to delay the deletion.
+				$delayed = get_post_meta( $attachment_id, Sync::META_KEYS['delay'], true );
+				$now     = time();
+				if ( empty( $delayed ) ) {
+					update_post_meta( $attachment_id, Sync::META_KEYS['delay'], $now );
+					$file = get_post_meta( $attachment_id, '_wp_attached_file', true );
+					update_post_meta( $attachment_id, '_' . md5( $file ), $file );
+				}
+				$valid = file_exists( get_attached_file( $attachment_id ) );
+			} else {
+				// Remove the delay meta.
+				delete_post_meta( $attachment_id, Sync::META_KEYS['delay'] );
 				$file = get_post_meta( $attachment_id, '_wp_attached_file', true );
-				update_post_meta( $attachment_id, '_' . md5( $file ), $file );
+				delete_post_meta( $attachment_id, '_' . md5( $file ), $file );
 			}
-			$valid = file_exists( get_attached_file( $attachment_id ) );
-		} else {
-			// Remove the delay meta.
-			delete_post_meta( $attachment_id, Sync::META_KEYS['delay'] );
-			$file = get_post_meta( $attachment_id, '_wp_attached_file', true );
-			delete_post_meta( $attachment_id, '_' . md5( $file ), $file );
 		}
 
 		return $valid;
+	}
+
+	/**
+	 * Validates the Stats mechanism.
+	 *
+	 * @param int $attachment_id The attachment ID to validate.
+	 *
+	 * @return bool
+	 */
+	public function validate_stats( $attachment_id ) {
+		return $this->delivery->is_deliverable( $attachment_id );
 	}
 
 	/**
@@ -490,6 +514,8 @@ class Storage implements Notice {
 		$this->download = $this->sync->managers['download'] ? $this->sync->managers['download'] : new Download_Sync( $this->plugin );
 
 		if ( $this->is_ready() ) {
+			$this->delivery = $this->plugin->get_component( 'delivery' );
+
 			$defaults       = array(
 				'offload' => 'dual_full',
 			);
@@ -507,6 +533,7 @@ class Storage implements Notice {
 
 			$structure = array(
 				'generate' => array( $this, 'size_signature' ),
+				'validate' => array( $this, 'validate_stats' ),
 				'priority' => 16,
 				'sync'     => array( $this, 'size_sync' ),
 				'state'    => 'info syncing',
@@ -524,5 +551,16 @@ class Storage implements Notice {
 				add_filter( 'wp_image_editors', array( $this, 'disable_editors_maybe' ) );
 			}
 		}
+	}
+
+	/**
+	 * Is storage set ot be full quality.
+	 *
+	 * @return bool
+	 */
+	public function is_local_full() {
+		$sync_media = $this->media->get_settings()->get_value( 'sync_media' );
+
+		return ! empty( $sync_media['offload'] ) && 'dual_full' === $sync_media['offload'];
 	}
 }
